Refactor - check 메서드를 util로 추출
